Revise build.xml and the search for Scala/ANTLR3
The previous ANT build.xml rebuilt all classes every time because the directory structure did not match the Java package/class names. This is troublesome not only because of the wasted time, but also 'sudo make install' would create files as root, making it difficult to delete them again.

Also supports out-of-tree builds now.

Previously, ANTLR3 and Scala files had to obey a very specific directory structure that does not correspond to the structure of most Linux distributions.  We are using the search functions from the autoconf-archive and use CLASSPATH if possible. The pencil-driver superproject cares for the correct versions of these projects if not installed on the system.